pyproject.toml
[tool.poetry]
name = "final_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
autoflake = "^1.3.1"
isort = "^4.3.21"
black = "^19.10b0"
unify = "^0.5"
flake8 = "^3.7.9"
pylint = "^2.4.4"
mypy = "^0.761"
fastapi = "^0.54.1"
uvicorn = "^0.11.3"
python-multipart = "^0.0.5"
pyjwt = "^1.7.1"
passlib = "^1.7.2"
sqlalchemy = "^1.3.16"
pytest-asyncio = "^0.10.0"
pytest-mock = "^3.0.0"
mock = "^4.0.2"
pillow = "^7.1.1"
aiofiles = "^0.5.0"
aioredis = "^1.3.1"
rq = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
